backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
beautifulsoup4==4.12.2
requests==2.31.0
weaviate-client==3.24.1
sentence-transformers==2.2.2
nltk==3.8.1
scikit-learn==1.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.4.2